Move portControl and linkControls to interfaces subdirectory.
