+++ src/sst/elements/merlin/Makefile.am
@@ -13,12 +13,6 @@ libmerlin_la_SOURCES = \
 	merlin.h \
 	merlin.cc \
 	router.h \
-	linkControl.h \
-	linkControl.cc \
-	portControl.h \
-	portControl.cc \
-	reorderLinkControl.h \
-	reorderLinkControl.cc \
 	bridge.h \
 	bridge.cc \
 	offeredload/offered_load.h \
@@ -66,6 +60,12 @@ libmerlin_la_SOURCES = \
 	inspectors/circuitCounter.cc \
 	inspectors/testInspector.cc \
 	inspectors/testInspector.h \
+	interfaces/linkControl.h \
+	interfaces/linkControl.cc \
+	interfaces/portControl.h \
+	interfaces/portControl.cc \
+	interfaces/reorderLinkControl.h \
+	interfaces/reorderLinkControl.cc \
 	arbitration/single_arb.h \
 	arbitration/single_lru.h \
 	arbitration/single_rr.h \
@@ -84,7 +84,6 @@ EXTRA_DIST = \
 
 sstdir = $(includedir)/sst/elements/merlin
 nobase_sst_HEADERS = \
-	linkControl.h \
 	router.h
 
 libmerlin_la_LDFLAGS = -module -avoid-version $(PYTHON_LDFLAGS)
