+++ src/sst/elements/merlin/hr_router/xbar_arb_age.h
@@ -26,7 +26,7 @@
 #include <queue>
 
 #include "sst/elements/merlin/router.h"
-#include "sst/elements/merlin/portControl.h"
+#include "sst/elements/merlin/interfaces/portControl.h"
 
 namespace SST {
 namespace Merlin {
