+++ src/sst/elements/merlin/hr_router/xbar_arb_lru_infx.h
@@ -25,7 +25,7 @@
 #include <vector>
 
 #include "sst/elements/merlin/router.h"
-#include "sst/elements/merlin/portControl.h"
+#include "sst/elements/merlin/interfaces/portControl.h"
 
 namespace SST {
 namespace Merlin {
